ITS FULL WEBSITE FOR BOOKING WEDDING HALL
The Wedding Hall Booking System is a web application designed to simplify the process of reserving and managing wedding hall bookings. The system provides an intuitive interface for users to browse available wedding halls, check their availability, and make bookings.

The frontend of the application will be developed using HTML, CSS, and JavaScript. HTML will be used to structure the web pages, CSS will handle the visual styling, and JavaScript will add interactivity to the user interface.

The backend of the application will be powered by PHP, which will handle the server-side logic and data management. It will interact with a database to store and retrieve information about wedding halls, bookings, and user details.

Users will be able to search for available wedding halls based on their preferred date, location, capacity, and other criteria. The search results will be dynamically displayed on the page using JavaScript.